+++ ext/opentelemetry-ext-pymysql/src/opentelemetry/ext/pymysql/__init__.py
@@ -13,8 +13,8 @@
 # limitations under the License.
 
 """
-The integration with PyMySQL supports the `PyMySQL`_ library and is specified
-to ``trace_integration`` using ``'PyMySQL'``.
+The integration with PyMySQL supports the `PyMySQL`_ library and can be enabled
+by using ``PyMySQLInstrumentor``.
 
 .. _PyMySQL: https://pypi.org/project/PyMySQL/
 
@@ -25,11 +25,13 @@ Usage
 
     import pymysql
     from opentelemetry import trace
-    from opentelemetry.ext.pymysql import trace_integration
+    from opentelemetry.ext.pymysql import PyMySQLInstrumentor
     from opentelemetry.sdk.trace import TracerProvider
 
     trace.set_tracer_provider(TracerProvider())
-    trace_integration()
+
+    PyMySQLInstrumentor().instrument()
+
     cnx = pymysql.connect(database="MySQL_Database")
     cursor = cnx.cursor()
     cursor.execute("INSERT INTO test (testField) VALUES (123)"
@@ -51,7 +53,7 @@ from opentelemetry.ext.pymysql.version import __version__
 from opentelemetry.trace import TracerProvider, get_tracer
 
 
-class PymysqlInstrumentor(BaseInstrumentor):
+class PyMySQLInstrumentor(BaseInstrumentor):
     def _instrument(self, **kwargs):
         """Integrate with the PyMySQL library.
         https://github.com/PyMySQL/PyMySQL/