+++ ext/opentelemetry-ext-pymysql/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Unreleased
 
+- Implement PyMySQL integration ([#504](https://github.com/open-telemetry/opentelemetry-python/pull/504))
 - Implement instrumentor interface ([#611](https://github.com/open-telemetry/opentelemetry-python/pull/611))